Allow for easy customization of Kotlin options in `kotlin-dsl` projects
This commits remove usage of `afterEvaluate {}` from the `kotlin-dsl`
plugin. This is possible now that it uses the Kotlin Gradle plugin
>= 1.8.0 that provides lazy configuration properties.

This allows for easy configuration of the KotlinCompile tasks directly,
thus Kotlin compiler options.

The standalone script compilation is also now configured to skip the
pre-release check in order to allow referencing Kotlin code compiled
with more recent Kotlin language versions on a best effort basis.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>