Make foreground daemon respect the instrumentation agent feature flag
Prior to this commit, the foreground daemon was working like a real
daemon process: if the agent is applied, it is used. However, it isn't
correct after we started to unconditionally apply the agent in the start
script. The foreground daemon starts to use the agent by default even if
the flag is actually not on-by-default.

This commit properly configures GlobalScopeServices created for the
foreground daemon by forwarding the desired agent status from the daemon
parameters.