Properly handle agent instrumentation feature flag in embedded runner
When starting a forked daemon with the embedded runner, agent should be
applied regardless of agent status of the runner itself. It is similar
to the start script where the agent is also applied regardless of the
feature flag. Without this it isn't possible to fork with "--no-daemon"
or "--foreground" with the embedded runner.