Postpone adding transformer to agent until we know it's needed
Initializing the transformer in `ProcessBootstrap` also adds the
instrumenting infrastructure to the client JVM (the one that talks to
the daemon and doesn't run builds). The overhead is actually noticeable
for the short builds (about 20ms extra).

This commit only sets up the transformer for the process that is going
to run the build, and only if the feature flag is enabled.