Rebaseline performance tests to accept regression introduced by agent
Applying the agent adds 10-20ms of startup time overhead to the JVM.
This also includes "client" JVMs that do not run builds, because the
start script (`bin/gradle`) has no way of figuring out if the
invocation is going to be client-only or to host a build.

In "quick" performance tests, the significant chunk of time is spent
starting up the client JVM, so the agent cost is visible on such test,
as a 1-2% regression.

Gradle needs the agent to be able to instrument signed JARs, so the
cost has to be accepted.

See gradle/gradle-private#3747