Rename JvmVariantBuilder to JvmFeature
The existing JvmVariantBuilder does a lot more than creating a variant. Its own documentation also seems
a bit confused about what it does, given that it claims to create an entire component when its implementation
clearly adds its domain objects to an existing component. The documentation has been updated to reflect the
actual current naming schemes we've been solidifying as part of the Idiomatic JVM effort.

In general, this PR is similar to the changes which resulted in the DefaultJvmSoftwareComponent. The
DefaultJvmVariantBuilder has been renamed to DefaultJvmFeature and has been updated to own the source
set that it wraps along with all domain objects it creates. The structure has also been updated to mirror
the internal structure of the DefaultJvmSoftwareComponent.

These changes brought to light the striking similarities between a 'JvmFeature' and the existing
'JvmSoftwareComponentInternal'. Given these similarities, there are future plans to replace most of
the logic in the jvm software component by instantiating 'main' jvm feature. Some additional shuffling
is needed to maintain backwards compatiblility, so this work will be done in a follow-up PR. The eventual
vision is that users will be able to register features directly on a component.

This PR also makes some minor changes to the JavaPlugin by moving some code which did not belong in the DefaultJvmSoftwareComponent
back into the JavaPlugin.