Provide a clear failure message for BuildService as configuration input
The typical way to make something an input to the configuration is to
use the ValueSource, and then obtain its value at configuration time.
Gradle recomputes these ValueSources when checking if the configuration
cache is still valid. However, at this point not all build state is
restored, but only a subset of it. This subset is not enough to restore
BuildServices if these are used as ValueSource inputs.

Before Gradle was failing the build with rather cryptic "cannot find
project :" error, and only when running from the cache - the initial
storing build was fine, even with load-after-store.

With this commit, a special set of codecs is used to store the
fingerprinting data, in which the BuildServiceProviders are explicitly
not supported. This way we can detect the providers even in long mapping
chains.

We may revisit the decision and make it possible to use BuildServices in
the future, but it requires a large rewrite of the fingerprinting code.