Clear root metadata when extending configurations
The hierarchy is saved as an instance variable in configuration metadata. The mutation type for configuration extension
was DEPENDENCIES, which only caused the cached metadata to be reevaluated. This is not sufficient. We add a new mutation
type for HIERARCHY which completely discards the metadata, allowing it to be recalculated and the new hierarchy value
to be passed to the configuration metadata

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>