Make failed ValueSources in the fingerprint a CC problem
If the ValueSource obtained at configuration time throws an exception,
then the configuration cache entry is discarded. The build itself may
not fail, if the exception is caught in the build script.

The failed value is still stored in the fingerprint, in case the user
runs build with `--configuration-cache-problems=warn`.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>