Handle failed ValueSources in the fingerprint at load time
Previously, we were get()-ting the fingerprinted value without checking,
and this caused the stored exception to be re-thrown. Now a throwing
ValueSource becomes a configuration cache problem, so we can expect that
only the users specifiyng "--configuration-cache-problems=warn" are
going to hit this, for everyone else the entry is discarded at store
time. Given this, it is less important to show a big red failure. The
user has no way out of this state other than removing the entry manually
so we help them and invalidate the entry with a short summary message of
what happened.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>