Rework the logic of adding new local variables in constructor instrumentation
Instead of the hacks with local variables sorter, read the maxLocals using the
ASM tree API and use the variables above the maxLocals value;

This requires reading classes with the ASM Tree API, but we can do it on-demand
only for the classes where we need that data. This can be further optimized to
building just the MethodNodes for methods that we need.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>