Place framework application dependenices at the beginning of the classpath
If we use a newer version of junit platform launcher which references new classes or fields in junit platform engine or commons,
we need to be sure that our loaded launcher sees those new class versions. Previously, we placed our loaded dependencies at then
end of the classpath, so if the testRuntimeClasspath has an old version of platform commons, our launcher would see classes from
that version. We move our loaded dependencies to the front of the classpath so that launcher sees its own version of
commons and engine.