Allow Configuration#extendsFrom to be called from within withDependencies
Make sure we run the dependency actions on the entire hierarchy before traversing that hierarchy
for dependencies and excludes. This is because the dependency actions may themselves modify the hierarchy,
so before traversing we need to make sure the hierarchy isn't going to change out from under us.

This should be fine since configurations' dependency actions are idempotent. The actions automatically
de-register themselves, so once they're called, they're not triggered for subsequent calls.