Fix reporting of successfull documentation tests
When a wrapping RunNotifier was first introduced, it only forwarded
information about failed tests, because only this method was overriden.
Successful tests weren't reported because methods of the original
notifier weren't called.

This commit adds a proper delegating RunNotifier that forwards all calls
emitted by SamplesRunner to the original RunNotifier, so no information
is lost.