Implement production logic on on top of Features
This reduces duplication and ensures the production logic and user-created features behave the same

Future work involves updating JVM software components to be multi-feature aware and dynamically advertise the variants of its features.
Furthermore, we should deprecate calling registerFeature without a proper JVM component present and update registerFeature to add its
features to the component to leverage the auto-variant-advertising functionality. Finally, we should remove the "extending" behavior
of registerFeature when using the main source set.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>