Clean up STABLE_CONFIGURATION_CACHE setup for process tests
We used to only complain about external processes when the feature is
enabled, but since the configuration cache became stable in 8.1 the
nagging is enabled regardless of the feature status.

This commit removes all setups of STABLE_CONFIGURATION_CACHE related to
the external process tests, as it is no longer needed.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>