Generate Kotlin extensions for Gradle API at build time
This removes the generated API jar for Kotlin extensions, saving a dozen of seconds on first use.
Instead, the same generation is invoked at build time when packaging the distribution.

This commits changes the build logic to use the code generation implementation from the wrapper.
Subsequent changes will move the code generation implementation to the build logic.