Introduce caching in LocalComponentInAnotherBuildProvider
The local `LocalComponentGraphResolveState`s for any component (project)
that is in "another build" is the same everywhere in the build tree.
It does not matter from which build in the tree the component is seen.
Therefore, we can cache these resolve states on the BuildTree scope
to be reused across all dependency resolutions of all builds.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>
Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>