Allow `notCompatibleWithConfigurationCache` tasks to configure others
Typical task configuration code calls `getProject` freely, and under
normal circumstances this is allowed with the configuration cache.
However, there are cases when tasks are configured at execution time, in
particular, when a task resolves a configuration that other tasks
contribute their outputs to. Resolving isn't allowed with the
configuration cache, but `notCompatibleWithConfigurationCache` provides
a escape hatch. This marking only supresses problems coming from the
marked task though.

Prior to this commit, `getProject` and related calls were always
attributed to tasks upon which the call was made. This commit also takes
into account the currently running task, so if the current task is marked
as `notCompatibleWithConfigurationCache` and causes some other tasks
to configure, then all calls to `getProject` and alike of these tasks
are also suppressed.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>