Revert classpath ordering of distribution-loaded test framework classes
When running embedded integration tests with test distribution, the junit-platform-launcher jar which is loaded from
the gradle distribution is different than the version we usually test against. This version is incompatible with
older versions of junit-platform-commons and junit-platform-engine jars pulled in by the corresponding junit jupiter versions.

To resolve this, we reversed the ordering of distribution loaded test framework jars to ensure our versions of these
junit-platform-launcher dependencies came first on the classpath. This turned out to break the newest version of junit jupiter,
which we do not test against, leading to a regression.

This reverses the order and ignores the embedded tests which break when testing against older versions of junit jupiter.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>