Reproducer test for kotlin script compilation cache misses
This test exercise kotlin script compilation avoidance when the
build script classpath contains JARs coming from an included
build for build logic.

In that case the JARs are not reproducible as the Jar task is
not cacheable and reproducible archives are not the default.

This causes unnecessary script recompilation.

This worked before moving to an agent based build logic classpath
instrumentation for the configuration cache because of a side
effect of the instrumentation: the patched JARs were made
reproducible.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>