Let kotlin script compilation avoidance fallback to a runtime classpath fingerprint
The fallback is triggered when encountering a Kotlin inline function
as these are not supported by our kotlin script compilation avoidance.

Before we did fallback to the fingerprint of the full JARs and in case
of unreproducible JARs this caused unnecessary script recompilations.

This commits changes this to fallback to the "runtime classpath" fingerprint
of such JARs, ignoring changing archive metadata.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>