Add variant visibility
Allow variants to declare if they are publised to external repositories. There are plenty of variants
which should not be published. For example, test results, sources for jacoco, etc. Variants could be
used to support local functionality as opposed to produce artifacts for consumption externally.

Also update test fixtures and registerFeature to add their features to the java component