Adds a current year binding, so that the copyright notice in the generated wrapper scripts can always be current
Note that this commit is the first of 2 needed to implement this change.  Need to get the new method into the API of a snapshot to update the current wrapper, to make this new method available to the build-logic in order to actually update the templates and make use of this new data.