Use ResolveContext instead of ConfigurationInternal in resolve pipeline
ResolveContext models 'something that can be resolved'. We update the resolution pipeline to use this type instead
of the more generalized 'ConfigurationInternal' so that in the future we may not need a full configuration to
perform a resolution.

This further allows us to start splitting the configuration type into separate types for resolvable, consumable, and declarable

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>