Add support for non-present providers in attribute containers
Attribute containers currently do not support providers which have no value. This limits us from leveraging
laziness in some places where we conditionally want to add attributes to a container. With this change,
we are able to get rid of the internal beforeLocking method on configurations by updating those
methods to adding attribute providers to the configuration's container.

We cannot remove the beforeLocking method entirely, since the gradle/gradle build uses that method, but
once we update the wrapper this should be possible.

When implementing this feature, I chose to make it so non-present attribute providers are not visible
when querying the attribute container. So, querying keySet, contains, etc will act as if the attribute
was never set on the container. I chose this approach since I believe it is a common use-case to first
check if an attribute is present on the container with a keySet or contains check, and then if so, call
getAttribute and assume the return-value is non-null.