Default to file.encdoing=UTF-8 on s390x arch
On an IBM Mainframe (s390x architecture) the default file.encoding of
the JVM is IBM-1048 (EBCDIC). Gradle expects the JVM to run in UTF-8.
Issues appear if either the client or the daemon JVM run in EBCDIC.

Instead of changing the code it makes most sense to make sure the JVM is
started in UTF-8 on s390x as well.

Resolves: #24498
Signed-off-by: David Frenzel <david.frenzel.crosswind@gmail.com>