Remove LenientResolutionResult
LenientResolutionResult was used to allow DependencyInsightReportTask to access non-fatal errors from a resolution, like
dependency locking and version conflict failures, and format them separately in the task output. However, implementing an
entire new ResolutionResult here was not necessary at all and could easily be accomplished without a new implementation.

This PR removes that implementation and simplifies the ResolutionResult codepath.