Add estimations of code size to LambdaSerializationTransformer
The generator needs to know beforehand if adding another deserialization
clause will exceed the method size budget. There is no obvious way of
calculating the potential size of the method with ASM before actually
emitting the code, so the estimation is hand-counted.

The added test verifies that the estimate matches the size of the
actually generated code.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>