Split generated $deserializeLambda$ if it becomes too large
Gradle makes all lambdas on the build script classpath serializable with
the bytecode instrumentation. A crucial part of it is generating the
synthetic $deserializeLambda$ method in the enclosing class that
instantiates lambdas from the serialized form. When the class has many
lambdas, the resulting method may exceed JVM size limit of 64Kb.

This commit adds method splitting: if the next lambda support will push
the method over the edge, then a new method for the remaining lambdas is
generated instead. The former method just calls to the new one as its
last operations.