Bump instrumentation version for lambda serialization improvements
While the primary goal is to fix instrumentation of JARs that cannot be
processed before, it is still possible that some classes are now
instrumented differently (e.g. if the class was at the verge of breaking
the limit, it may now get a second split method because estimation is
over-estimating).