Add a future-proofing check for lambdas that cannot be deserialized
These cannot exist in practice now because of the JVM limits and how
lambdas are implemented. However, if we start to generate more code for
deserialization of if JVM allows more arguments, we may see this check
failing. Without it we'll see infinite loops that are a pain to debug.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>