Add smoke test for gradle-consistent-versions
Adds a smoke test for gradle-consistent-versions, which has unconventional code that
at the very least, copies configurations, consumes them via dependency-management, and
abuses beforeResolve to extend other configurations.

We have broken this plugin before, so even though it is not a major pulgin in the ecosystem,
its unconventional methods are useful to determine if we are breaking behavior. In fact, adding
this smoke test brought up two existing regressions which broke this plugin:

This PR fixes these regressions
1. Previously copies of configurations with locked usages would have a locked usage themselves. This
is different behavior than in 8.0 where copies would have unlocked usages. This was an 8.1 regression
2. This plugin manages to add dependencies to the consumable-only apiElements configuration. We had
an assertion which verified the configuration had no dependencies. This PR removes that assertion,
though we should follow-up here and determine how this plugin managed to add dependencies to apiElements.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>