Fix ClassCastException in Groovy DSL plugins { ... } block syntax checker (#7897)
The code unconditionally casted the arguments Expression of the MethodCallExpression
to ArgumentListExpression. But it could also have other types like for example
NamedArgumentListExpression when the arguments are given as map. This happens
if you have a legacy plugin apply within the plugins { ... } block and then
caused a cryptic Groovy compiler error instead of a clear message what is wrong.

Signed-off-by: Björn Kautler <Bjoern@Kautler.net>
Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>