Canonicalize "." and ".." segments in RelativePaths
Fixes #5748

As documented in #5748, Sync and Copy tasks can have buggy
behaviors (such as deleting copied files) when RelativePath objects
refer to the same path but consist of a different list of segments.
This can happen in particular when "." and ".." are used in path
segments.

Avoid this situation by canonicalizing RelativePaths as they are
created, so that "." is not used and ".." only appears at the
beginning of a path.

Signed-off-by: Alex Landau <alandau@cs.stanford.edu>