Refactor TaskExecutionMode to hold an arbitrary message
TaskExecutionMode was turned into an Interface and DefaultTaskExecutionMode was created to implement it.
DefaultTaskExecutionMode provides facotyr methods for each of the types of execution modes that was previously the enum constants.
DefaultTaskExecutionMode also can accept a reason for `untracked` execution modes that can be used to enrich the execution context with the reason of the track not being tracked

https: //github.com/gradle/ge/issues/20936
Signed-off-by: Szava Maczika <smaczika@gradle.com>
Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>