Re-enable flaky tests with explicit timeout
The Tooling API cross-version tests verifying download progress events
were flaky on CI. Sometimes the execution hung and the build failed
with a timeout. Consequently, the tests were marked as @Flaky.

The problem is not reproducible locally with the data available. On
top of that, the existing reports don't have enough details about the
failing execution: the timeout generates a thread dump from the build
agent and not from the test execution process.

The tests were marked as @Flaky quite a while ago, so there's also a
chance that the problem no longer exists. Test execution data from GE
seems to confirm that.

This PR tentatively re-enable the flaky tests and defines an explicit
timeout on them. After merging, if we don't see flaky tests then removing
the @Flaky annotation was indeed correct. If the failures re-appear,
we'll then get stacktraces from the test worker and can continue with
the investigation.