Remove flakiness annotation to check to capture data
This test was marked as flaky, but enough time has passed that we no longer have any data
about the flaky occurrences.  We don't don't have any occurrences in the last 90 days and
attempte to reproduce the flakiness have not succeeded.  If it re-occurs, let's capture
the occurrence and re-annotate until we can fix it.