Remove vendor specific JVM implementations
These are not adding much value and can cause issues when the logic ages and newer
versions don't get classified properly.  Instead, pull the apple-specific environment
variable filters up into the standard JVM as these are unlikely to be important anywhere
else and there are reports that these are added on the J9 JDK, too.  We still need to
distinguish IBM from non-IBM JDKs, so we'll keep that logic around.