Create ArtifactViews using "live" attributes from creating configuration
This commit adds a new test class, ArtifactViewAttributesIntegrationTest, which ensures that (assuming no additional configuration) the attributes an ArtifactView uses in selecting files and artifacts are identical to those used by the configuration that created it.

Certain attributes are adding to Java configurations by the beforeLocking callback by the JavaBasePlugin and DefaultJvmPluginServices classes.  Previously, if an ArtifactView was created prior to these callbacks being triggered, the attributes they added were not present in the AV's collection when the AttributeMatchingVariantSelector selected a variant to obtain files from the ArtifactView.

Depending on the order files were accessed, this could cause AV's with no additional configuration to return a different set of files than the configurations that created them.

Instead of copying an immutable collection of attributes, just use the existing collection of attributes, as we defensively copy the attributes of the VA if they are requested to be configured.