Deprecate DefaultArtifactPublicationSet
This is a stepping-stone to deprecating the archives configuration. We simplify the DefaultArtifactPublicationSet to simplly
pass-through all candiates to the backing archives configuration.

Future changes to deprecate the archives configuration will update the configuration to emit deprecation warnings when adding
artifacts to it. Instead, users should explicitly add task dependenices to the assemble task