Use "live" attributes when fetching artifacts from configurations
Certain attributes are adding to Java configurations by the beforeLocking callback by the JavaBasePlugin and DefaultJvmPluginServices classes.  Previously, if artifacts was accessed prior to these callbacks being triggered, the attributes they added were not used to retrieve artifacts.

Instead of copying an immutable collection of attributes, just use the existing collection of attributes.