Avoid fileCollection(Spec) in the gradle build
The previous implementation used the filtred file collection, selecting for all external dependenices
and their transitive dependencies.

The relatively simple solution of replacing this with selecting for all external dependencies regardless
of which dependencies they are reachable from _works_, but it degrades performance. This is because, when
we actually go to load these modules from the distribution at runtime, we end up wasting time deduplicating
the external modules that all local project dependencies all have in common.

So, instead as a replacement, we walk the graph ourselves during build-time to discover all external
dependencies that our project dependencies would not otherwise pull in transitively. This is leads to
us loading an even smaller subset of external dependencies per distribution module than before, potentially
leading to a performance improvement over the past. This is with the downside of a more complex build, however
we want to avoid the loss of performance by using the basic 'include all external deps' solution.