Polish code quality plugins
The AbstractCodeQualityPlugin presented poor abstractions to the child code quality plugins
and as such led to duplication and unnecessary code. Particularly, all child implementations
were jvm-specific and the abstract plugin was also JVM specific, but there were abstractions
which avoided applying the java base plugin by default. This was unnecessary and complicated.

Further changes updated the abstract plugin to avoid exclude rules in favor of artifact views
and component fiters. Furthermore, resolvable configurations which intended to resolve JVM
artifacts were updated to be configured as runtime classpaths.

Finally, a public method on PmdPlugin was deprecated for removal in 9.0 since it does not
need to be public.