Simplifiy bridge between TaskContainer and Software Model
This bridge uses several internal APIs specific to the TaskContainer. In the future we should consolidate these.
This PR merges two classes which are highly coupled into one and updates them to use lambdas instead of anonymous
inner classes. This change makes it easier to reason about the usages of these specialized internal methods
since they are all localized in one file instead of two.