Create initial public Configuration factory methods
It is the goal of the JVM team to ensure that no configuration can change roles starting in Gradle 9.0.
There are a few reasons for this. One being that it is difficult to apply conventions to all configurations
of one type if the role of that configuration can change after the conventions are applied. The other more
important reason is that it is difficult to have a single implementation for a Configuration which can
fulfil all three usages while also limiting memory usage and complexity. This may be possible with some sort
of state machine and delegation to usage-specific implementations, however this would likely be complicated to
implement.

To resolve this, we introduce new APIs on ConfigurationContainer to create configurations in a pre-defined role:
1. Resolvable 2. Consumable 3. Dependencies.
These configurations created with these factory methods will not be able to change roles -- contrary to existing
factory methods where after initial creation the configuration can change to any role as desired.

Future work will deprecate creating configurations with the standard create and register methods and will also
deprecate changing the role of any configuration after it has been created.

At a high level, this PR makes the following changes:
- Creates new public factory methods for all 3 usages in ConfigurationContainer
- Defines the ConsumableConfiguration, ResolvableConfiguration, and Dependencies Configuration marker interfaces
    - The factory methods return these types. Users can use instanceof to check the type of a configuration
    - Future work will add toX methods on these interfaces when we introduce new APIs for consumable, resolvable, dependencySpec
- Introduce 'unlocked' versions of public factory methods for internal use in RoleBasedConfigurationContainerInternal
    - These methods create non-role-locked versions of the public methods. In 9.0 we will transition to the public APIs, but cannot do so yet to avoid breaking changes
- Add default implementations for the configuration-extending interfaces marked above. Update ConfigurationFactory to be able to instantiate any of these
- Make DefaultConfiguration abstract. Future PRs will likeley update the name of this but for now to avoid potential breaking changes this will remain with the existing name
- Remove existing internal methods for creating configurations with arbitrary roles, exposing only internal methods to create configurations in known non-deprecated roles or deprecated roles marked in ConfigurationRolesForMigration
- Update Ear plugin to no longer create legacy configurations. 'deploy' and 'earlib' will be transitioned to resolvable dependency configurations
- Update internal plugins throughout to create unlocked configurations using new internal factory methods