Avoid beforeLocking
beforeLocking is a bad way to achieve laziness. It is only used to add attributes to an attribute container, which itself supports Providers. This PR updates our usages of beforeLocking to use providers and deprecates the beforeLocking method. Eventually, once the beforeLocking method is removed entirely, we will have more freedom to cleanup the mutation situation for Configurations -- particularly, we will not need to lock all configurations for mutation when a single configuration is resolved.

Before we can remove beforeLocking, we need to remove this usage of it in the gradle/gradle build logic -- though that may prove difficult