Fix resolution of excluded tasks for included builds
When a project with an included build, where the included build in turn
includes a build for a convention plugin, Gradle will try to resolve
excluded tasks before the main project is fully created when trying
to configure convention plugin project.

This change fixes the behavior of the resolveExcludedTaskName to return
an empty filter if any of the build's projects have not yet created,
the same that already occurs when the project is not yet loaded.

Resolves: #24341
See also: #21936

Signed-off-by: Tyler King <tylerking0@gmail.com>