Merge DefaultPendingSource into SortedSetElementSource
When a pending element is realized, the ElementSource notifies the container with the realized value, which then via onRealize
calls BACK to the ElementSource with the same element. This cycle is confusing. In order to simplify this and remove the cycle,
we need to make sure SortedSetElementSource is 'aware' of its pending elements and does not need to rely on onRealize to update
the realized elements.

To do this, we first need to merge the pending part of SortedSetElementSource with the realized part. We do this by merging
DefaultPendingSource, which is only used by SortedSetElementSource into SortedSetElementSource. This also makes SortedSetElementSource
more similar to AbstractIterationOrderRetainingElementSource