Mitigate change in behavior
With the previous implementation of DefaultDomainObjectCollection's addLater and addAllLater,
upon adding a new provider, if the container was subscribed to elements of the provier's type,
the container would call 'doAddRealized'. This would then subsequently tell the ElementSource
the _realized_ versions of all values in the provider, without telling the ElementSource which
provider those values came from.

Now, with these changes, the provider is aware of which provider the values came from. The previous
behavior allowed immediately realized providers to be lazily evaluated _once_ before their provider
values were cached. While this behavior may seem favorable, it is actually a result of a bug where
the ElementSource was not aware of the fact that the backing provider was already realized, and as
such did not cache its output.

In the long run, we should remove this hack by making the ElementSources more aware of changing
providers and thus they will be able to clear their own caches once the source of the data changes
as opposed to needing to eagerly clear their own caches