Create initial JVM platform
Creates the initial platform directory structure by adding a `platforms` directory at the repo root, next to `subprojects`.
Inside this dir is a directory for each platform. This PR creates the JVM platform directory, named `jvm`.
We then moved the `:ear` project to that directory as a proof-of-concept. The physical moving of the ear files is done in the
next commit.

The build is still not "platform aware". `subprojects.json` was updated to include these new platforms in their flattened form,
so that there is no conceptual difference between declaring a subproject in `subprojects` and in `platforms/<platform>`.

This paves the way for other teams to start taking ownership of their subprojects and moving them into their corresponding
platform directories. Future work will involve making the build more platform-aware, such as adding ArchTest rules for ensuring
certain platforms do not access projects from other platforms. For example, the execution platform should not need to access anything
from the JVM platform.