Remove as much internal usage of ResolvedConfiguation as possible
This removes internal usage and test usage as much as is possible.  When alternative to ResolvedConfiguration are NOT present when testing multiple ways of resolving, the ResolvedConfiguration is replaced with ArtifactViews.  When use of ArtifactView was already present, these tests were left as is - when RC is removed, they should become simple case deletions.