Allow Configurations with only outgoing attributes to be selected
Configurations with no attributes cannot be selected via variant aware dependency management.
When checking for configurations with no attributes, we only consider the attributes defined on the base configuration.
This PR updates the check to also check for attributes declared on ConfigurationPublications, since those attributes
are also added to configuration metadata during selection