Extract ResolutionStrategy Factory
Create a standalone ResolutionStrategyFactory. This logic should be separate from a ConfigurationContainer.

The primary motivation for this was to avoid calling 'startParameterResolutionOverride.applyToCachePolicy(cachePolicy)'
during resolution, as this mutatates the resolution strategy. Ideally, we should not be mutating the resolution
strategy _after_ resolution has been initiated.

Making this change centralizes the logic required to build a resolution strategy. But more importantly,
by avoiding mutating the resolution strategy after resolution has been initiated, we can leverage
this change to be more strict about when we lock configurations for mutation