Add new KMP smoke test and refactor Kotlin smoke tests
This refactors the class hierarchy of the existing Kotlin tests to make the purpose of each class more clear.  It also reduces duplication by adding new abstract base classes and traits to remove duplicate code and static methods across class heirarchies.  It avoids needless re-parsing of version number using the VersionNumber class when checking to add deprecation warning expectations to these tests.  It updates the usage of the Kotlin workers API parameter to test not only true/false values but also absence, which is the current recommendation.

It also adds a new smoke test for a JS KMP project based on a sample from the Kotlin docs.