Make CodeNarc output more user friendly
When console output is enabled, use the IDE formatter instead of the text formatter.
The IDE formatter produces clickable links to files and line numbers.

When html report is enabled, use the sortable formatter instead of the html formatter.
The sortable formatter produces a HTML report with buttons to sort failures by rule name,
rule priority, most violations or file.

When multiple reports are enabled, we sometimes disabled the link to a report in the failure
message. This is because we only looked if the first report was not the console. We now look
for any report that's not the console. We prefer linking to HTML > text > XML.