Don't calculate fork point commit on security fork
In the past, we calculated the "fork point commit" as performance baseline
on feature branches. However, in security fork, because the forked
repository is private, `git fetch` and subsequent `git merge-base`
command would fail, resulting in the calculated performance baseline
to be `8.2-commit-`.

This PR checks if the current repostory is a security fork (gradle/gradle-ghsa)
and skip calculating fork point commit in this situation.