Deprecate getTaskDependencyFromProjectDependency
This method is by definition incompatible with project isolation. We will remove it publicly in 9.0
and will transition away from it internally, where it is used for the buildNeeded and buildDependents
tasks when the Java plugin is applied. We will need to figure out other ways to implement this sort
of functionality. Though, efficiently implementing buildNeeded is likely impossible