Filter unsupported versioned directories when instrumenting MRJARs
ASM can only process class files up to a certain version. This can be
problematic for multi-release JARs containing newer classes in the
appropriate versioned directory. Such JARs can be used on a
Gradle-supported JVM, but prior to this change, instrumentation failed
even if the unsupported classes were never loaded.

This commit makes instrumentation skip unsupported versioned
directories. For legacy instrumentation, these directories are filtered
out completely. For agent-based instrumentation, classes of the
unsupported versions are not instrumented and aren't placed into the
transformed archive.

Unfortunately, ASM doesn't provide any kind of constant to specify the
latest supported class file version, so we have to keep track of this
manually.