Omit link to CC report when no CC errors exist
A link to a CC report is considered distracting if there are no CC
problems - even more so if the build is already
failing for non-CC reasons. This change set no longer includes a link
to the CC report if no CC errors occur (no matter whether there are
other non-CC failures). The link is still shown with info logging level,
or if an internal option is provided (useful for CC tests).

Fixes: #24248