Allow lazy configuration dependencies to resolve current project
Prior to 8.2, addAllLater hooks on a configuration could resolve other configurations in the same project in order
to determine the dependencies to add. 8.2 added a change which eager calculated confguration metadata dependenices
for all consumable variants in the same project, causing a stackoverflow. This once again makes configuration dependency
calculation lazy, avoiding a stackoverflow. Note, a stackoverflow will likely still happen if the configuration with the
lazy dependencies selects itself. But, this makes no sense conceptually.