Disable persistent compiler daemons in ForkCapableRelocationIntegrationTest
On Windows, the persistent compiler daemon holds on to the java agent jar
that is kept in the build directory.

This prevents these tests from running clean or cleaning up after the tests.

For now, disable persistent compiler daemons when this happens in this test.

In general, we need some way to detect that a persistent compiler daemon is
depending on files in an "unsafe" location and either isolate those files
or stop the compiler daemon before we need to delete things.