Disable embedded testing jetty server stop timeout
Currently, the embedded jetty server used in integration tests uses the default 30s stop timeout.
Jetty waits for half the stop timeout time before shutting down when a stop request is sent.
This causes test cases which use this test fixture to take at least 15 seconds to execute.
As a result, some tests take much longer than necessary, reducing testing turn-around time, for example
when running all tests in a given subproject, even if only one of those tests use the fixture.
'DependencyUnresolvedModuleIntegrationTest' is a prime example of this, and is often the last
test to remain executing when running all dependency-management tests.