Recognize certain options as making workers unreliable on Windows
When arguments like -javaagent are used with a worker, its possible
for this option to add a file that is held open by the worker across
builds.

If this is a jar that comes from the dependency cache, this is mostly
OK.

If this is a jar that comes from another project, this breaks simple
things like running clean because we're unable to delete the build
directory without stopping the worker process.

This change causes the daemon option builder to switch to SESSION-persistent
daemons. This has been the default for a long time and mostly works.

This change doesn't eliminate all possible avenues to cause this problem.
There are other ways to pass options to the worker process that
could cause it to hold on to files (even non-jars).