Allow java-test-fixtures to be applied standalone
Previously, java-test-fixtures requried the Java plugin to be applied to add any new behaviors.
Instead, we now perform _most_ of the work without requiring the Java plugin to be applied, but
instead automatically apply the JavaBasePlugin.

We still require react to the Java plugin in order to add automatic dependencies between the
test fixtures and the production code, and the tests and the test fixtures. We also react to the
Java plugin in order to publish the test fixtures. However, both of these operations are possible
to perform manually without the Java plugin being applied.